Extracts base class for tries 📦
